code cleanup for #4 -- renamed IndexCreationStatus.java to IndexCreatedEvent.java to better fit naming conventions; deprecated IndexCreatedHandler; better failures for indexing taking too long
