+++ integration_test/src/test/java/com/strategicgains/docussandra/controller/IndexControllerTest.java
@@ -229,7 +229,7 @@ public class IndexControllerTest
             f.insertDocuments(docs);//put in a ton of data directly into the db
 
             Index lastname = Fixtures.createTestPlayersIndexLastName();
-            
+
             String tableStr = "{" + "\"fields\" : [\"" + lastname.fields().get(0)
                     + "\"]," + "\"name\" : \"" + lastname.name() + "\"}";
             RestAssured.basePath = "/" + lastname.databaseName() + "/" + lastname.tableName() + "/indexes";
@@ -292,7 +292,7 @@ public class IndexControllerTest
                     break;
                 }
                 LOGGER.debug("Waiting for index to go active for: " + sw.getTime());
-                if (sw.getTime() == 12000)
+                if (sw.getTime() >= 12000)
                 {
                     fail("Index took too long to create");
                 }
@@ -305,7 +305,7 @@ public class IndexControllerTest
             RestAssured.basePath = restAssuredBasePath;
         }
     }
-    
+
     /**
      * Tests that the POST /{databases}/{table}/indexes/ endpoint properly
      * creates a index and that the GET/{database}/{table}/index_status/
@@ -352,15 +352,15 @@ public class IndexControllerTest
 
             //check to make sure it shows as present at least once
             ResponseOptions res = expect().statusCode(200)
-                    .body("_embedded.indexcreationstatus[0].id", notNullValue())
-                    .body("_embedded.indexcreationstatus[0].dateStarted", notNullValue())
-                    .body("_embedded.indexcreationstatus[0].statusLastUpdatedAt", notNullValue())
-                    .body("_embedded.indexcreationstatus[0].eta", notNullValue())
-                    .body("_embedded.indexcreationstatus[0].index", notNullValue())
-                    .body("_embedded.indexcreationstatus[0].index.active", equalTo(false))//should not yet be active
-                    .body("_embedded.indexcreationstatus[0].totalRecords", notNullValue())
-                    .body("_embedded.indexcreationstatus[0].recordsCompleted", notNullValue())
-                    .body("_embedded.indexcreationstatus[0].precentComplete", notNullValue())
+                    .body("_embedded.indexcreatedevents[0].id", notNullValue())
+                    .body("_embedded.indexcreatedevents[0].dateStarted", notNullValue())
+                    .body("_embedded.indexcreatedevents[0].statusLastUpdatedAt", notNullValue())
+                    .body("_embedded.indexcreatedevents[0].eta", notNullValue())
+                    .body("_embedded.indexcreatedevents[0].index", notNullValue())
+                    .body("_embedded.indexcreatedevents[0].index.active", equalTo(false))//should not yet be active
+                    .body("_embedded.indexcreatedevents[0].totalRecords", notNullValue())
+                    .body("_embedded.indexcreatedevents[0].recordsCompleted", notNullValue())
+                    .body("_embedded.indexcreatedevents[0].precentComplete", notNullValue())
                     .when().get("/").andReturn();
             LOGGER.debug("Status Response: " + res.getBody().prettyPrint());
             //wait for it to dissapear (meaning it's gone active)
@@ -372,8 +372,8 @@ public class IndexControllerTest
                 LOGGER.debug("Status Response: " + body);
 
                 JSONObject bodyObject = (JSONObject) parser.parse(body);
-                JSONObject embedded = (JSONObject)bodyObject.get("_embedded");
-                JSONArray resultSet = (JSONArray)embedded.get("indexcreationstatus");
+                JSONObject embedded = (JSONObject) bodyObject.get("_embedded");
+                JSONArray resultSet = (JSONArray) embedded.get("indexcreatedevents");
                 if (resultSet.isEmpty())
                 {
                     active = true;
@@ -381,9 +381,9 @@ public class IndexControllerTest
                     break;
                 }
                 LOGGER.debug("Waiting for index to go active for: " + sw.getTime());
-                if (sw.getTime() == 12000)
+                if (sw.getTime() >= 12000)
                 {
-                    fail("Index took too long to create");
+                    fail("Index took too long to create: " + body);
                 }
                 Thread.sleep(5000);
             }
@@ -394,7 +394,6 @@ public class IndexControllerTest
             RestAssured.basePath = restAssuredBasePath;
         }
     }
-
 
     /**
      * Tests that the DELETE /{databases}/{table}/indexes/{index} endpoint
